Fix - Remove additional file
